+++ src/test/clojure/clara/test_fressian.clj
@@ -17,15 +17,15 @@
   (with-open [os (java.io.ByteArrayOutputStream.)
               ^FressianWriter wtr (fres/create-writer os :handlers df/write-handler-lookup)]
     ;; Write
-    (pform/thread-local-binding [d/node-id->node-cache (volatile! {})
-                                 d/clj-struct-holder (java.util.IdentityHashMap.)]
-                                (fres/write-object wtr x))
+    (pform/with-thread-local-binding [d/node-id->node-cache (volatile! {})
+                                      d/clj-struct-holder (java.util.IdentityHashMap.)]
+      (fres/write-object wtr x))
     ;; Read
     (let [data (.toByteArray os)]
-      (pform/thread-local-binding [d/clj-struct-holder (java.util.ArrayList.)]
-                                  (with-open [is (java.io.ByteArrayInputStream. data)
-                                              ^FressianReader rdr (fres/create-reader is :handlers df/read-handler-lookup)]
-                                    (fres/read-object rdr))))))
+      (pform/with-thread-local-binding [d/clj-struct-holder (java.util.ArrayList.)]
+        (with-open [is (java.io.ByteArrayInputStream. data)
+                    ^FressianReader rdr (fres/create-reader is :handlers df/read-handler-lookup)]
+          (fres/read-object rdr))))))
 
 (defn serde [x]
   ;; Tests all serialization cases in a way that SerDe's 2 times to show that the serialization to